The year is 2018- All things have a right to grow.
It's time to end plant racism. And better yet, pretend it never existed in the first place even though the commit logs clearly show otherwise. Grasses are great in their own right, so long as you're not trying to grow wetland grasses in a desert. In fact, they provide food for lots of animals which can help defend the area's trees. They can help absorb floods and do other great things. Overall, lawns consisting of nothing but grass are terrible ideas (well a million times better than astroturf lawns but still not amazing), but when planted in the understory this grass is generally positive.